pyproject.toml
[project]
name = "energy_drink_rating"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.123.4",
    "numpy>=2.4.1",
    "passlib[bcrypt]>=1.7.4",
    "psutil>=7.1.3",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
]
